requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.5
django-allauth==0.44.0
django-crispy==1.8.1
django-import-export==2.5.0
django-markdown-deux==1.0.5
django-quill-editor==0.1.16
et-xmlfile==1.0.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib-metadata==3.4.0
jdcal==1.4.1
Markdown==3.3.3
markdown2==2.3.10
MarkupPy==1.14
moviepy==1.0.3
numpy==1.19.5
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.1.0
proglog==0.1.9
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
tablib==3.0.0
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.26.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
